ranking previewをsuspenseしてみた
