+++ src/main.tsx
@@ -1,17 +1,19 @@
-import React from "react";
+import React, { Suspense } from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import AppRoutes from "./App";
 import "./index.css";
 import Layout from "./layout";
+import { LoadingPage } from "./pages/Loading";
 
-// biome-ignore lint/style/noNonNullAssertion: <explanation>
 ReactDOM.createRoot(document.getElementById("root")!).render(
-	<React.StrictMode>
-		<BrowserRouter>
-			<Layout>
-				<AppRoutes />
-			</Layout>
-		</BrowserRouter>
-	</React.StrictMode>,
+  <React.StrictMode>
+    <BrowserRouter>
+      <Suspense fallback={<LoadingPage />}>
+        <Layout>
+          <AppRoutes />
+        </Layout>
+      </Suspense>
+    </BrowserRouter>
+  </React.StrictMode>
 );
